fix typo

+++ Src/Newtonsoft.Json.Tests/JsonTextReaderTests/ReadAsyncTests.cs
@@ -513,18 +513,18 @@ namespace Newtonsoft.Json.Tests.JsonTextReaderTests
 
             await jsonReader.ReadAsInt64Async();
             Assert.AreEqual(JsonToken.Integer, jsonReader.TokenType);
-            Assert.AreEqual(typeof(int), jsonReader.ValueType);
-            Assert.AreEqual(250, jsonReader.Value);
+            Assert.AreEqual(typeof(long), jsonReader.ValueType);
+            Assert.AreEqual(250L, jsonReader.Value);
 
             await jsonReader.ReadAsInt64Async();
             Assert.AreEqual(JsonToken.Integer, jsonReader.TokenType);
-            Assert.AreEqual(typeof(int), jsonReader.ValueType);
-            Assert.AreEqual(250, jsonReader.Value);
+            Assert.AreEqual(typeof(long), jsonReader.ValueType);
+            Assert.AreEqual(250L, jsonReader.Value);
 
             await jsonReader.ReadAsInt64Async();
             Assert.AreEqual(JsonToken.Integer, jsonReader.TokenType);
-            Assert.AreEqual(typeof(int), jsonReader.ValueType);
-            Assert.AreEqual(250, jsonReader.Value);
+            Assert.AreEqual(typeof(long), jsonReader.ValueType);
+            Assert.AreEqual(250L, jsonReader.Value);
 
             Assert.IsTrue(await jsonReader.ReadAsync());
             Assert.AreEqual(JsonToken.EndArray, jsonReader.TokenType);
